separate application js to movement_category.js
